Move stat and chmod after all vmcore extraction and stripping is done.
The previous attempt at fixing the following bug was incorrect:
https://bugzilla.redhat.com/show_bug.cgi?id=1076833

Move the stat and chmod to the very end, after all vmcore extraction
and stripping (makedumpfile) processing is done.  The previous patch
did not catch the case where makedumpfile was run, and also messed
up the log message which calculated the space saved by makedumpfile
(it always said "Makedumpfile ... saved 0 bytes" after the last change).

Moving the stat and the chmod to the very end of all processing is
where it should be.
Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1076833

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>
Signed-off-by: Michal Toman <mtoman@redhat.com>