Convert crash_cmd from string to mutable list type inside prepare_debuginfo
Sometimes running a default 'crash' command fails due to various reasons.
In at least some instances, we can recognize the failure and add some further crash
options which work better and allow crash to run and not exit with an error.
In order to do this and utilize the 'crash_cmd' member of RetraceTask, we need
to either:
a) move all code which calls any 'crash' function inside RetraceTask
b) use a mutable type, such as list, instead of immutable string type

Choose option 'b' above since option 'a' really does not make sense as it would
require large code refactorings which probably are not necessary.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>