Fix incorrect group permissions when writing RetraceServer files especially with interactive mode
When users run either retrace-server-interact or retrace-server-worker directly, these commands
run as uid:gid.  This may result in incorrect gid ownership on files in a task's directory.
In some cases, as in the 'crash_cmd' file, incorrect group ownership can result in a second user
being unable to run interactive mode.

Fix the above by doing a chgrp to CONFIG["AuthGroup"] everytime we write any file in the
RetraceTask directory.  This patch only addresses the gid part, not any issue with the uid or
the mode bits.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>