Fix incorrect group permissions when writing RetraceServer files especially with interactive mode
When users run either retrace-server-interact or retrace-server-worker directly, these commands
run as uid:gid.  This may result in incorrect gid ownership on files in a task's directory.
In some cases, as in the 'crash_cmd' file, incorrect group ownership can result in a second user
being unable to run interactive mode.

Fix the above by doing a chgrp to CONFIG["AuthGroup"] everytime we write any file in the
RetraceTask directory.  This patch only addresses the gid part, not any issue with the uid or
the mode bits.  Make the chgrp a "best effort" and so ignore any error on the chgrp which
may occur if we don't own the file.  If we don't own the file, then probably we don't need
to chgrp anyway, but it's cheaper to just ignore the error than check if we own the file first,
etc.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>