Enforce uniform mode bits for almost all RetraceTask files
With interactive mode there's a real danger that users can run with a restrictive
umask, locking out other users, and even leading to a full outage of the 'manager',
if for example the 'status' file is unreadable by both group and other.  Specifically
if a user does
$ umask 0077
$ retrace-server-worker --restart <taskid>
It will cause the manager page to fail to load and the user will just get "Internal Error"

Fix all this by enforcing uniform permissions on all files as:
-rw-r--r--.

We make one exception and that is the 'crash_cmd' file.  We leave this as group
writeable in case something comes up where we need another crash commandline
parameter for some vmcore, or in case something goes wrong and we need to reset
it.

Unfortunately there are still a couple exceptions which are subject to the user's
umask, namely retrace_log and password, but these don't cause major problems.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>