Add vmlinux file inside RetraceTask
Make the path to the vmlinux file persistent by adding in a file which stores
the path to the vmlinux file to be used with crash.  This file is extracted
during kernel version detection.

In later patches we will need this file so we can avoid calling prepare_debuginfo
(and rpm -ql kernel-debuginfo) in a redundant way.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>