Move prepare_debuginfo and strip_vmcore inside RetraceTask, avoid prepare_debuginfo from retrace-server-interact
We should not have to detect anything when a user invokes retrace-server-interact.
The kernel version and vmlinux file should already be setup properly prior to this.
In order to do this, we have to move prepare_debuginfo and strip_vmcore inside RetraceTask.
Also, at the return points of prepare_debuginfo, write out the path to the vmlinux file.

Existing tasks may not have a vmlinux file though, so check for existence, and if no
vmlinux file exists, call prepare_debuginfo as before to set it up.

This change results in a significant performance boost when loading any vmcore.
In my tests, I measured anything from 5-10 seconds between a load where the vmlinux file
does not exist and one where it does (and we skip prepare_debuginfo).  All my tests
were on an idle system too so the boost is probably greater in production.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>