Avoid calling prepare_debuginfo from retrace-server-interact after kernel version detection.
In the normal case, when a user invokes retrace-server-interact, we should immediately be able
to invoke crash on the vmcore and vmlinux file, and should not need to detect anything.
In the retrace-server-worker, the kernel version should have been detected, and the
vmlinux file should have already been extracted from kernel-debuginfo.

Prior patches moved prepare_debuginfo and strip_vmcore into RetraceTask and introduced
vmlinux file into RetraceTask.  With these prior patches, we now check for existence
of the vmlinux file, and if it exists, we can avoid calling prepare_debuginfo.

One special case we have to handle though is for existing tasks.  Existing tasks prior to
this release will not have a vmlinux file.  To handle this, check for existence of it, and
if it does not exist, we must call prepare_debuginfo as before to set it up.

This change results in a significant performance boost when loading any vmcore.
In my tests, I measured anything from 5-10 seconds between a load where the vmlinux file
does not exist (we need to call prepare_debuginfo) and one where it does exit (and we skip
prepare_debuginfo).  All my tests were on an idle system too so the boost is probably
greater in production.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>