stats: Update the release information to be dynamic based off of plugins.
Prior to this update, the stats page regarding releases would simply return a count
for the Fedora 15-17 releases as queried from the backing database. Following this
update, the process for gathering this data is more dynamic in nature.

The individual plugins now include a display string and individual release entry
list similar to the following:

    versionlist = [
      "fc1",
      "fc2",
      "fc3",
      "fc4",
      "fc5",
      "fc6",
      "fc7",
      "fc8",
      "fc9",
      "fc10",
      "fc11",
      "fc12",
      "fc13",
      "fc14",
      "fc15",
      "fc16",
      "fc17",
      "fc18",
      "fc19",
      "fc20",
      "fc21",
      "fc22",
      "fc23",
      "fc24",
    ]

As a result, the release list is now dynamic and will note the specific release
for the list above in each of the defined plugins.

Signed-off-by: Kyle Walker <walker.kyle.t@gmail.com>