manager.wsgi: Move the FTP query operation to an AJAX operation
In the event that UseFTPTasks is enabled, and the backing FTP server
contains an exceptional number of files in it, the resulting query
can take an extreme amount of time to return. In the meantime, the
manager UI is unavailable for use to access existing tasks.

This patch implements an AJAX call to a new "ftp" interface. It will
return the remainder of the "FTP files" table by default, but will
also adhere to the "filter=" variable passing to limit the resulting
table.

Signed-off-by: Kyle Walker <walker.kyle.t@gmail.com>