Make the reposync tool more verbose if required
Sometimes someone does something bad to repo files and directories and
rsync's output seems to be the best source for fixing such problems.

This commit adds possibility to pass more '-v' on command line. If there
are more 2 or more '-v's, then output of executed commands (rsync,
createrepo) is forwarded to stdout/stderr; otherwise the output goes to
/dev/null.

Signed-off-by: Jakub Filak <jfilak@redhat.com>