Correct eu-unstrip parser if FILE is -
When 'eu-unstrip -n' can find the matching ELF and debug files on
disk given the build-ids it will display them, otherwise it won't.
If it can find the files then it "resolves" the module name on the end,
otherwise the name of the module is file path.

Example:
    0x7f248b+0x20d 96030f@0x7f248 - - /usr/lib64/gedit/plugins/libtime.so

If the FILE part does not begin with '/' and the MODULE part is not
'[exe]', coredump2packages silently skips these eu-unstrip output
lines ignoring the fact that such lines are completely valid.

This commit ensures that those lines will be processed too.

Signed-off-by: Jakub Filak <jfilak@redhat.com>