Fix duplicate email if a vmcore task fails to determine the kernel version.
The only caller of start_vmcore is wrapped inside a try ... except statement.
We thus don't need to explicitly call '_fail' method here but can raise an exception.

This resolves a duplicate email as a result of calling _fail inside start_vmcore
then again in the try ... except clause inside the start method.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>