Not mark packages with different architectures as duplicity
Packages with the same name, epoch, version and release but with different
architectures should not be considered the same package. One of them will be
deleted and it is not clear which one.

It happened that these 3 packages were found after processing output from
eu-unstrip:
   1) python-debuginfo-2.7.12-5.fc24.x86_64
   2) python-2.7.12-5.fc24.i686
   3) python-2.7.12-5.fc24.x86_64
No.2 and no.3 were marked as duplicity.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>