Correct eu-unstrip parser if FILE is .
When an ELF image is available but not from any named file, 'eu-unstrip -n'
sets FILE to '.' and MODULENAME is set to a file path.
Lines where FILE is not "-" or does not begin with '/' are skipped.
This eliminates valid lines.

This commit process also these lines.

When FILE is "."?
    From help to eu-unstrip: "FILE is the file name found for the module,
                              or - if none was found, or . if an ELF image is
                              available but not from any named file."
    My view: If there is a path in a coredump, but this path cannot be found
             on disk, then FILE is . and MODULNAME is the path.

Example I got from eu-untrip on real coredump:

0x55a47e199000+0x202000 5d36a3439f5b7bc5e321e73f3647c3c3187156f3@0x55a47e199284 . - /usr/bin/python2.7
0x7ffe937d9000+0x1000 cd24458eaca84d42b76906559c1e564f1ad2222f@0x7ffe937d97d8 . - linux-vdso.so.1

Signed-off-by: Matej Marusak <mmarusak@redhat.com>