Enable creating repository from faf repository
In faf there was introduced a way of assigning operating system with release
to builds. That makes it possible to find all builds of specific operating
system and release. Then repository can be created by using packages from faf
repository.

Action of creating repository from faf was integrated into
retrace-server-reposync, so when this action is executed and UseFafPackages
in configuration file is enabled, repository is created/updated from faf.

This closes abrt/retrace-server#115.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>