Use devtoolset-4-gdb when used on RHEL
There main reason why using gdb from devtoolset is a good idea is that, there
is a newer version of gdb present in devtoolset-4. New gdb improves the way, in
which backtraces are produced and therefore overall quality of backtrace can be
higher and more readable.
Another reason why new gdb is needed, is presence of argument `-ascending`
which can print thread is ascending order. This adds on readability of
backtrace from multithread application.

To make this commit possible, two main things had to made:
1) Plugins must keep information, which gdb version do they need. They also
need to keep information about the way in which their version of gdb can be run.
These information are later used when mock is created and when gdb is run.

2) After founding release for specific task, RetraceWorker class have to save
a found plugin, so later in the same class, this plugin can be easily accessed.
This also meant moving guess_release method into this class, co it can set this
attribute directly.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>