Use short form rhel when creating repository
In the faf database, name of rhel is saved in the full form "Red Hat Enterprise
Linux". Therefore to the reposync-faf script this form must be passed.
This was inconvienient when this script was called from retrace-server-reposync,
since the output repository would be also called
"Red Hat Enterprise Linux-XX-AAAA".
Now the opsys can be passed as rhel and it will be found inside fafs database.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>