Include md5sum of original archive in summary page
This commit adds option for users to see md5sum of original archive. User can
select before submitting task whether this information will be displayed or not.

If selected, this information is displayed on summary page as well as is written
into file. Format of this file is same as when md5sum command is run, so these
data can be compared easily (for example with diff).

Also into configuration file new item was added, to have this checkbox checked
by default.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>