Cleanup tmp FAF repository after failing
The process of retracing can fail. That meant that sometimes tmpdir
created from faf-c2p was not removed.

The _fail method was called at least twice, between creating and
deleting this tmpdir. Also a _fail method could be called from different
function that was called between creating and deleting tmpdir.
Therefore I've decided to put fafrepo as class attribute, so it is
possible to delete from anywhere.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>