Set kernelver and vmlinux as soon as possible
For larger vmcores it is possible that prepare_debuginfo may take a long time.
This is due to various factors such as makedumpfile, as well as how long crash
takes to load when detecting modules as well as running the early commands
such as 'sys'.  Today we are not setting kernelver or vmlinux as early as we
can and as a result, someone waiting to get access to the vmcore either has to
wait excessivly long or do some manual intervention like setting up vmlinux
manually.  Today if someone tries to run 'retrace-server-interact' before all
commands complete in the retrace, they will end up blocking waiting
for the re-detction of the kernelver which can take a very long time (since
it may run strings on the whole file).  This makes no sense since retrace
figures out the kernelver early on as well as the vmlinux file (assuming it
is in the cache).  This patch sets both the kernelver file and the vmlinux
file as early as reasonably possible, and thus allows much earlier access
to the vmcore through 'retrace-server-interact <task> crash'.

Resolves https://bugzilla.redhat.com/show_bug.cgi?id=1516329

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>