Convert notify_email_success and notify_email_fail to a single method
There's a lot of duplicated code, so create a single notify_email
method and use the task.get_status() method to change the output
slightly.

Note there is a couple small functional changes.  First, in the
'_fail' function we need to call task.set_finished_time prior to
notify_email.  Second, in the successful case we need to set
STATUS_SUCCESS before calling notify_email.  Finally we fix a small
typo in the success case (originally the subject had "succeded").

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>