Add md5sum and kernelver to email notifications, help text to failing notification
This patch adds md5sum and kernelver to success or fail email notifications.
In addition, if a kernel task fails, add helper text to point the user in the
direction of possibly remedies.

For now, we just add some generic text which may or may not apply to a given
failure, but is probably the most common failure.  Future patches could expand
the task status values to different failure values, and we could revisit the
helper text based on specific failures.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>