Remove duplicate RetraceWorker._fail call when start_vmcore fails
The only place start_vmcore is called from is the start method of
RetraceWorker, and that call is inside a try ... except clause which
calls _fail in the 'except' part.  Thus, inside start_vmcore, if we
have a nested exception, we don't need to call _fail there.  Instead
we should just raise an exception and let the 'start' method call
_fail with our Exception string set to the describe the failure.

This patch also adds an error message if mock is being used and
prepare_debuginfo or one of the crash commands fail.  This makes
it consistent with the non-mock case.  Once we move this except
clause un-indent the rest of the block to make it consistent.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>