Remove duplicate RetraceWorker._fail call when start_vmcore fails
The only place start_vmcore is called from is the start method of
RetraceWorker, and that call is inside a try ... except clause which
calls _fail in the 'except' part.  Thus, inside start_vmcore, if we
have a nested exception, we don't need to call _fail there.  Instead
we should just raise an exception and let the 'start' method call
_fail with our Exception string set to the descript the failure.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>