Improve vmcore kernel parsing for certain scenarios and limit file scanning
The existing parser has a couple downsides:

1. It may scan the whole file with 'strings' if it cannot find the kernel version.
This causes the OS page cache to be constantly polluted with useless pages
and can impact other vmcore processing, especially if a huge file is submitted
that fails detection.  This covered in
https://bugzilla.redhat.com/show_bug.cgi?id=1440381

2. It can be faked out by other package versions that match the kernel
version parser.  This problem is covered in
https://bugzilla.redhat.com/show_bug.cgi?id=1535592

This patch fixes both of the above bugs.

The new code keeps the initial use of "crash --osrelease" since in testing
we see this correctly identifies the kernel version on 60% of vmcores.

For the remaining 40% of vmcores, first only scan the first portion of the file
(we use 64MB based on empirical evidence of thousands of vmcores) and do not
use "strings" anymore on the whole file.  Then use a couple regexes to search
for the kernel version.  This method of reading the file and then going through
regex's one by one is superior to reading one line from the 'strings' output
in that it allows us to use the less error-prone regex's first, then
only when these fail, go to the ones that are more error prone.

For the first regex we keep the "OSRELEASE=" search which is fairly reliable
to identify kernel versions (we've seen no instances where this shows up for
a non-kernel version).  However, note that "crash --osrelease" may do
'strings -n 10' on the file looking for "OSRELEASE=", so in one sense this
regex is redundant.  However, in our testing keeping this did improve detection
for some older vmcores, and it does not cost much, so we keep the regex.

The second regex is a new one based on a longer string that identifies a
kernel version along with the build date and is more unique than just a
kernel version regex.  This helps with the problem in bug 1535592.

The final regex is to fall back to a small variation on our original
KERNEL_RELEASE_PARSER, adjusted only for the new method of reading the file
(we are no longer reading strings of the file but reading the file directly
so the terminator is not the end of line).  Note that this final regex is
subject to finding other non-kernel package versions, so it's not ideal.
Even still it does provide some value from our testing so we keep it.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>