Clean up handling of mock with x86 vmcores in x86_64 environments
Originally when RetraceTask.use_mock was introducted it was done
alongside the implementation of RetraceTask.crash_cmd.  When crash_cmd
was introduced it was not well understood where we really should set
this and read it.  Also, use_mock was not really well done either.

The right thing to do with crash_cmd and use_mock is only set both of
them once the arch has been properly detected from get_kernel_release.
Create methods get_mock(), set_mock(), and has_mock() to handle whether
mock is configured or not.

This cleans up the code quite a bit and fixes on bug which occurs
when 32-bit cores are run on a 64-bit machine where prepare_debuginfo
would be called twice:
https://bugzilla.redhat.com/show_bug.cgi?id=1536683

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>