Set default signal handler for SIGPIPE before calling Popen on 'crash --osrelease'
In testing another bug involving rework of get_kernel_version() it was found that
for some vmcores calling Popen("crash --osrelease"...) would CPU spin for a while
before returning a valid result.  Oddly if you run from a shell 'crash --osrelease'
on the same vmcore it does not CPU spin but returns in less than 1 second.  This
is described in https://bugzilla.redhat.com/show_bug.cgi?id=1540253#c0

As it turns out, there's a python bug where SIGPIPE is not set to the default
handler and this is what causes 'crash --osrelease' to CPU spin.  The fix is
fairly straightforward - just set the handler to default prior to calling Popen.
A general fix inside Popen was rejected for RHEL6 due to side-effects so here
we only do a fix limited to the affected Popen command.

We could revisit a more comprehensive fix but it would require further scrutiny
and more comprehensive testing.  For now just patch this one Popen that we know
is affected and set the handler back to what it was prior to the Popen.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>