Fix typo in exception handling of get_kernel_release
The previous commit 232ff36 introduced a minor bug when processing
a vmcore that failed a permission check or read error.  The retrace_log
will show a message indicating 'IOerror' not being defined.
For example:
 2018-02-28 15:17:51 prepare_debuginfo failed: global name 'IOerror' is not defined
 2018-02-28 15:17:51 File '/retrace/tasks/875129564/crash/vmcore' is not group readable and chmod failed. The process will continue but if it fails this is the likely cause.

While the overall error message is still accurate, there is a small bug
here.  Indeed "IOerror" is not defined and this is a simple typo and this
should be 'OSError'.

Fixing this shows a slightly more accurate retrace_log:
 2018-02-28 15:24:38 prepare_debuginfo failed: [Errno 13] Permission denied: '/retrace/tasks/722567843/crash/vmcore'
 2018-02-28 15:24:38 File '/retrace/tasks/722567843/crash/vmcore' is not group readable and chmod failed. The process will continue but if it fails this is the likely cause.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>