Set md5sum as soon as possible.
The md5sum is calculated as soon as a file is downloaded but it is not
written to the file until after the task is processed.  As far as I
can tell the reason for this is due to the fact that multiple files
can in theory be inside the 'downloaded' file.  However, this leads to
the side-effect that a failed task may never write out the md5sum file
with the actual md5sum despite the fact it has been calculated.
Instead the only thing that remains in the file is "Enabled" which is
obviously not desireable.

Fix this by writing the md5sum as soon as it is calculated.  This may
lead to the file being written multiple times but this should not be
a problem.  Also in many instances (for example with vmcores) there is
only one file so it will only be written once.

NOTE: This problem was more obvious after the dedup_vmcore patch went
in.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>