Implement retrace-server-bugzilla-refresh tool
The tool goes through tasks with the bugzilla field filled in and checks the status
of entered bugzillas. If the task has bugzilla bug that is still open it is refreshed
to prevent deletion by the cleanup tool.

Tasks without the bugzilla field or tasks with only resolved bugzillas should then be
cleaned by retrace-server-cleanup tool.

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>