py3 compatibility: Replace filter function with a list equivalent
In Python 2, `filter(function, iterable)` returns a list and is equivalent
to `[item for item in iterable if function(item)]` if function is not `None`
and `[item for item in iterable if item]` if function is `None`. In Python 3,
the function returns an iterator. The replacement of Python 2 `filter()`
function with a proper equivalent or wrapping a call to `filter()`
with a call to `list()` solves the Python 3 compatibility.

Signed-off-by: Jan Beran <jberan@redhat.com>