py3-comp: Specify encoding for Popen
If encoding or errors are specified, or text is true, the file objects stdin, stdout and stderr
are opened in text mode with the specified encoding and errors. The universal_newlines argument
is equivalent to text and is provided for backwards compatibility. [1]

By default, file objects are opened in binary mode.

encoding argument was added in version 3.6.

[1] https://docs.python.org/3/library/subprocess.html#popen-constructor

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>