Fix get_kernel_release() on Python3
The get_kernel_release() function first calls 'crash --osrelease' and
if that returns a valid string it will use this.  At this point we have
the correct type.  However, if this fails, we then read binary data
from the vmcore file and use some regexp's to find kernel versions.
For Python3 we need the regexps to be byte objects.  Once we've found
some match, we're down to a small character set and so decode that into
a string.  Finally at the end we strip out any any spaces, newline or
return characters before returning the release.

This fixes retrace-server task failures with log entries that look like
this:
2019-03-15 14:09:37 prepare_debuginfo failed: cannot use a string pattern on a bytes-like object
    2019-03-15 14:09:37 Stripped size: 32.01 GB
    2019-03-15 14:09:37 Makedumpfile took 0 seconds and saved 0.00 B
    2019-03-15 14:09:37 Analyzing crash data
    2019-03-15 14:09:37 cannot use a string pattern on a bytes-like object

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>