Fix ftp.wsgi for Python3 due to sorted() method change, speed manager load
In Python3 the sorted() method changed to no longer have a 'cmp' method.
Further, the 'cmp' method is stated to be slow, and we've had issues
where the 'manager' page can be slow to load.  Per Python2 reference:
https://docs.python.org/2/library/functions.html#sorted
"In general, the key and reverse conversion processes are much
faster than specifying an equivalent cmp function. This is because
cmp is called multiple times for each list element while key and
reverse touch each element only once. Use functools.cmp_to_key()
to convert an old-style cmp function to a key function."

There is not a high value to any sorting or 'key' method on the manager
page since there's often many files, and users can just use CTRL-F to
find their file, click on it, and submit it.  Removing 'cmp' completely
has the side benefit that will speed manager page loading.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>