Refactor crash commands run after prepare_debuginfo into run_crash_cmdline helper
When processing a vmcore and after prepare_debuginfo() runs successfully,
we run a series of crash commands.  These commands are the same whether
mock is used or not.  Create a helper function, run_crash_cmdline which
removes a lot of redundancy and improve maintainability.  In addition
to refactoring and maintenance improvements, there are a couple slight
functional changes with this patch.

First, the patch changes what happens with crash commands when the
'if child.wait()' condition is true.  Prior to this patch, we would call
log_warn() then set the output to 'None', with the only exception being
the 'log' command which we would not set output to 'None'.  With this
patch, we never set the output to 'None' for any command.  This should
not be an issue for most commands, and might even be viewed as a feature
since it will create output files in 'misc' for these other commands,
possibly with just error output in them.  However there is one possible
issue with creating error output in these files and involves 'foreach_bt'.
In the case of error output in 'foreach_bt', probably we should skip
calling bt_filter.  To avoid calling bt_filter, add a len() check similar
to what we do with the kernellog - if it is less than 1024 bytes we
assume it is not valid and skip calling bt_filter.

Second, the new function should handle non-printable characters in the
output of any crash command.  Prior to this patch, a partially damaged
vmcore with one or more corrupted task stacks could exit with a cryptic
error and task failure:
2019-06-10 10:43:42 'utf-8' codec can't decode byte 0xac in position 10472: invalid start byte
This only started happening on Python3 (previous python2 code handles
the same vmcore fine and does not fail the task).  This patch handles
this case by wrapping the child.communicate() in a try...except and
printing the error in that case:
 2019-06-11 17:57:51 crash command: 'set hex; foreach bt; quit; ' triggered UnicodeDecodeError
 2019-06-11 17:57:51   'utf-8' codec can't decode byte 0xac in position 10522: invalid start byte
If we hit the above error, we do not obtain any output in the command that
failed, so it does slightly change the behavior from Python2.  However,
at least now the specific command is now flagged as having an error,
the task is not failed, and the user may decide what to do with the error.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>