Remove 'utf-8' encoding for run_crash_cmd and change related functions.
Use binary mode to write crash output to various files in 'misc'.  This
avoids decoding problems as may happen if a vmcore is damaged and
crash outputs non-'utf-8' characters.

Inside run_crash_cmd once we have the binary output, add a try .. except
around crash_cmd.decode('utf-8') so UnicodeDecodeError will flag any
crash output that contains unexpected (non-'utf-8') characters.
Any such output is likely a sign the vmcore is damaged or incomplete,
or some issue with crash tool, so add a log_warn if this occurs.

This change requires adding 'mode' option to a few functions, as well
as a small change to manager.wsgi which may read 'misc' files.  When
reading these files and sending in response, we have to decode the
bytes into string and ignore any non-'utf-8' bytes to avoid backtraces
and "Internal Errors" in the browser.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>